requirements.txt
tensorflow==2.18.0
numpy==1.26.4
opencv-python
matplotlib
pillow==11.2.1
scikit-learn
flask
huggingface_hub
gunicorn 